backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
python-dotenv
chromadb
pandas